feat(game): show a GM-called dice roll while the turn is still running
A roll the GM asked for now drives the same animated card a /roll shows,
instead of arriving as a number in the prose with no die behind it. The
tool_result event carries the parsed roll past the debug-only gate, so the card
appears mid-turn rather than after the message saves.

A saved roll no longer swallows its message. The card stood in for the whole
message body, which is right for a /roll - the message is nothing but the roll
- and wrong for a GM turn that has narration of its own. It now replaces the
content only for a non-assistant message and sits alongside it otherwise. No
message shipping today is an assistant turn carrying a roll, so nothing
currently rendered changes.

isDiceRollResult moves to lib so a hook can narrow an SSE payload without
importing the card component.
